Struct veryfast::pool::Pool

pub struct Pool<TYPE> {
    // some fields omitted
}

A fast heap-allocator. Allocates objects in a batch, but transfers the control to the Object

Methods

impl<TYPE> Pool<TYPE>

fn new(align_to_cache: bool, batch: usize) -> Pool<TYPE>

Creates a new Pool.

align_to_cache: Should each object be on a separate CPU cache line. Speeds up multithreaded usage but requires a bit more memory in most cases.

batch: How many objects should be allocated each time. Higher numbers are faster, but can cause wasted memory if too little are actually used.

fn add(&self, obj: TYPE) -> Object<TYPE>

Save the object on the heap. Will get a pointer that will drop it's content when dropped (like a Box). The memory will be reused though!

Thread-safe. Very fast most of the time, but will take a bit longer if need to allocate more objects.